ADD: Add handler to get pagination result

+++ src/main/java/com/anasdidi/ecommerce/common/PaginationDTO.java
@@ -9,7 +9,7 @@ import lombok.Value;
 @EqualsAndHashCode(callSuper = false)
 public class PaginationDTO extends BaseDTO {
 
-  private final Integer totalElements;
+  private final Long totalElements;
   private final Integer totalPages;
   private final Integer pageNumber;
   private final Integer pageSize;
